requirements.txt
about-time==4.2.1
absl-py==1.4.0
aiofiles==23.2.1
alive-progress==3.1.5
annotated-types==0.7.0
anyio==3.6.1
appdirs==1.4.4
appnope==0.1.4
argcomplete==1.10.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arviz==0.15.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
Backtesting==0.3.3
beautifulsoup4==4.12.3
bertopic==0.16.3
bitarray==2.9.2
bitstring==4.2.3
bleach==6.1.0
blis==0.7.11
bokeh==3.0.2
bs4==0.0.1
cachetools==5.2.1
cairocffi==1.7.1
CairoSVG==2.7.1
catalogue==2.0.10
certifi==2024.8.30
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.4
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
colorlover==0.3.0
comm==0.2.2
compressed-rtf==1.0.6
confection==0.1.5
configobj==5.0.8
configparser==6.0.1
constants==0.6.0
contourpy==1.0.6
cryptography==42.0.8
cssselect2==0.7.0
cvxopt==1.2.7
cvxpy==1.1.15
cycler==0.10.0
cymem==2.0.8
DateTime==4.5
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.6
docx2txt==0.8
ebcdic==1.1.1
ecdsa==0.19.0
ecos==2.0.7.post1
eikon==1.1.15
emcee==3.1.4
emoji==2.13.2
entrypoints==0.4
esptool==4.7.0
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.1
executing==2.0.1
extract-msg==0.28.7
fastjsonschema==2.19.1
filelock==3.13.1
finnhub-python==2.4.15
fix-yahoo-finance==0.1.37
flatbuffers==23.1.4
forex-python==1.6
fqdn==1.5.1
frontend==0.0.3
frozendict==2.4.6
fsspec==2024.9.0
future==1.0.0
gast==0.4.0
gEconpy==1.1.0
gitdb==4.0.11
GitPython==3.1.43
glob2==0.7
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grapheme==0.6.0
grpcio==1.51.1
h11==0.14.0
h2==3.2.0
h5netcdf==1.1.0
h5py==3.7.0
hdbscan==0.8.38.post1
hpack==3.0.0
html5lib==1.1
htmltools==0.6.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.25.0
hyperframe==5.2.0
idna==2.10
IMAPClient==2.1.0
importlib-metadata==6.0.0
install==1.3.5
intelhex==2.3.0
ipykernel==6.29.4
ipython==8.18.0
ipywidgets==8.1.3
iso8601==2.1.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.0.1
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
keras==2.11.0
kiwisolver==1.3.2
langcodes==3.4.0
language_data==1.2.0
libclang==15.0.6.1
linkify-it-py==2.0.3
llvmlite==0.39.1
looseversion==1.3.0
lxml==5.3.0
marisa-trie==1.2.0
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.0.2
MouseInfo==0.1.3
mpmath==1.2.1
multiprocess==0.70.14
multitasking==0.0.9
murmurhash==1.0.10
mysql-connector==2.2.9
narwhals==1.13.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.5
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
notebook==7.2.0
notebook_shim==0.2.4
numba==0.56.4
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
openpyxl==3.0.8
opt-einsum==3.3.0
Optimisation==1.2
orjson==3.10.11
osqp==0.6.2.post0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.3
pandas-datareader==0.10.0
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pathlib_abc==0.1.1
pathos==0.3.0
pathy==0.11.0
patsy==0.5.2
pdfminer.six==20231228
pdfplumber==0.11.4
peewee==3.17.7
pexpect==4.9.0
pickle-mixin==1.0.2
pillow==10.4.0
platformdirs==4.2.0
plot-metric==0.0.6
plotly==5.11.0
plotting==0.0.7
polars==1.31.0
pox==0.3.2
ppft==1.7.6.6
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.36
protobuf==3.19.6
prov==2.0.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydot==2.0.0
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pymssql==2.2.5
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
pynare==0.1.5
pynndescent==0.5.13
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyodbc==4.0.34
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.9.0
pyportfolioopt==1.4.2
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.12
python-pptx==0.6.23
pytweening==1.2.0
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0
pyzmq==24.0.1
qdldl==0.1.5.post0
QuantLib==1.26
questionary==2.0.1
rdflib==7.0.0
reedsolo==1.7.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.8.1
ro-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ro_core_news_sm-3.6.0/ro_core_news_sm-3.6.0-py3-none-any.whl#sha256=c50bc556c3c2f6770e8488037f70748cf6a617781d1c35bf27065708261d54a6
ro-legal-fl==3.6.1
rpds-py==0.18.1
rsa==4.9
rubicon-objc==0.4.9
safetensors==0.4.5
scikit-learn==0.24.2
scipy==1.10.1
scs==2.1.4
seaborn==0.11.2
selenium==4.11.0
Send2Trash==1.8.3
sentence-transformers==3.1.1
serial==0.0.97
shellingham==1.5.4
shiny==1.2.1
shinyswatch==0.8.0
shinywidgets==0.3.3
simplejson==3.17.5
six==1.12.0
sklearn==0.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.2.1
sovai==0.1.3
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-stanza==1.0.4
SpeechRecognition==3.8.1
srsly==2.4.8
stack-data==0.6.3
stanza==1.6.1
starlette==0.37.1
statsmodels==0.13.1
sympy==1.11.1
tenacity==8.1.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
terminado==0.18.1
textract==1.6.5
thinc==8.1.12
threadpoolctl==2.2.0
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
torch==1.13.1
tornado==6.2
tqdm==4.66.4
traitlets==5.14.3
traits==6.3.2
transformers==4.44.2
trio==0.27.0
trio-websocket==0.11.1
typer==0.9.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uc-micro-py==1.0.3
umap-learn==0.5.6
Unidecode==1.3.8
uri-template==1.3.0
urllib3==1.26.6
uvicorn==0.27.1
wasabi==1.1.3
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==2.2.2
widgetsnbextension==4.0.11
workbook==1.1
wrapt==1.14.1
wsproto==1.2.0
xarray==2023.2.0
xarray-einstats==0.5.1
xlrd==1.2.0
XlsxWriter==3.0.1
xlutils==2.0.0
xlwt==1.3.0
xyzservices==2022.9.0
yfinance==0.2.46
zenpy==2.0.25
zipp==3.11.0
zope.interface==5.4.0
